REFACTOR: 폴더/파일명 및 구조 변경(#95 이슈 참고)
