+++ frontend/src/pages/RequestForm.js
@@ -1,7 +1,7 @@
 import Button from "components/Button";
 import Header from "components/Header";
 import { useState } from "react";
-import { buttonStyles } from "styles/props";
+import { buttonStyles } from "lib/styles";
 
 function RequestForm() {
   const [request, setRequest] = useState({ title: "", content: "" });
